+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minimum Rust version to 1.23.0
 - `Capstone::disasm()` methods take `&mut self` instead of `&self`
 
+### Removed
+- Duplicate/unneeded `Capstone` methods that have equivalents in `InsnDetail`
+    - `insn_belongs_to_group()`, `insn_group_ids()`, `register_id_is_read()`, `read_register_ids()`,
+      `register_id_is_written()`, `write_register_ids()`
+
 ## [0.3.1] - 2018-03-26
 ### Fixed
 - Documentation URL